Migrate Territory to core model
