+++ core/models.py
@@ -1,3 +1,12 @@
 from django.db import models
 
-# Create your models here.
+
+class Territory(models.Model):
+    """Categorical geographic region/territory"""
+
+    name = models.CharField(max_length=100, unique=True)
+    label = models.CharField(max_length=10, unique=True)
+    created_at = models.DateTimeField(auto_now_add=True)
+
+    def __str__(self):
+        return f"{self.name} - {self.label}"
