+++ pss/admin.py
@@ -1,7 +1,6 @@
 from django.contrib import admin
 
-from pss.models import Contact, Customer, Territory
+from pss.models import Contact, Customer
 
 admin.site.register(Contact)
 admin.site.register(Customer)
-admin.site.register(Territory)
