+++ pss/models.py
@@ -59,14 +59,3 @@ class Contact(models.Model):
 
     def __str__(self):
         return f"{self.customer.name} - {self.name}"
-
-
-class Territory(models.Model):
-    """Categorical geographic region/territory"""
-
-    name = models.CharField(max_length=100, unique=True)
-    label = models.CharField(max_length=10, unique=True)
-    created_at = models.DateTimeField(auto_now_add=True)
-
-    def __str__(self):
-        return f"{self.name} - {self.label}"
